app/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary
python-jose 
passlib[bcrypt]
pytesseract
python-multipart 
requests 
pytesseract 
pillow
alembic
python-dotenv